align design with figma
